fix: 修复评分功能失败及刷新后评分消失问题
核心问题：推荐API返回数据缺少id字段，导致前端BookCard发送book_id=undefined，后端返回422错误

修复内容：
1. 修改RecommendationItem schema添加id、predicted_rating、user_rating、avg_rating字段
2. 更新所有推荐引擎(CFEngine、SVDEngine、ColdStartHandler、GNNRecommender)返回完整数据
3. 修复CompareView.vue动态加载失败(computed未导入)
4. 修复InvalidCharacterError(el-radio的value属性问题)
5. 修复登录注册功能问题
6. 优化推荐引擎启动预加载，避免首次请求超时
7. 明亮主题全面覆盖(11+个前端文件)

技术改进：
- 后端增加调试日志便于问题排查
- 前端增加请求数据类型打印
- 评分API路径添加尾部斜杠避免307重定向
- RatingResponse的created_at设为可选避免序列化错误

+++ book-v2/backend/app/models/rating.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, DateTime, CheckConstraint
+from sqlalchemy import Column, Integer, Float, ForeignKey, DateTime, CheckConstraint
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.database import Base
@@ -10,7 +10,7 @@ class Rating(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     book_id = Column(Integer, ForeignKey("books.id", ondelete="CASCADE"), nullable=False, index=True)
-    rating = Column(Integer, nullable=False)
+    rating = Column(Float, nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
